+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Fixed Last words & trails for non op players
 * Fixed particle issues
 * Fixed invisible bug for spectators that joined while ingame
+* Fixed trail particle are random even on bow shoot  
 * Changed config.yml Item values of states and added all states
 
 ### 1.7.5 Release (01.02.2021 - 14.03.2021)